:memo: Adds README.md
- :fire: Also removes unnecessary `float64` conversion

+++ data/entities/receipt.go
@@ -29,12 +29,12 @@ func CountPoints(r *Receipt) int {
 		}
 	}
 
-	if math.Mod(float64(r.Total), 1.0) < 0.01 {
+	if math.Mod(r.Total, 1.0) < 0.01 {
 		// 50 points if the total is a round dollar amount with no cents.
 		points += 50
 	}
 
-	if math.Mod(float64(r.Total), 0.25) < 0.01 {
+	if math.Mod(r.Total, 0.25) < 0.01 {
 		// 25 points if the total is a multiple of 0.25.
 		points += 25
 	}
@@ -49,7 +49,7 @@ func CountPoints(r *Receipt) int {
 			// If the trimmed length of the item description is a multiple of 3,
 			// multiply the price by 0.2 and round up to the nearest integer.
 			// The result is the number of points earned.
-			points += int(math.Ceil(float64(item.Price) * 0.2))
+			points += int(math.Ceil(item.Price * 0.2))
 		}
 	}
